adds ViewModel for GameFragment
