+++ GuessTheWord-Starter/app/src/main/java/com/example/android/guesstheword/screens/game/GameViewModel.kt
@@ -1,0 +1,72 @@
+package com.example.android.guesstheword.screens.game
+
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.ViewModel
+
+class GameViewModel : ViewModel() {
+
+    // The current word
+    var word: String = ""
+
+    // The current score
+    var score = 0
+
+    // The list of words - the front of the list is the next word to guess
+    private lateinit var wordList: MutableList<String>
+
+    init {
+        resetList()
+        nextWord()
+    }
+
+    /**
+     * Resets the list of words and randomizes the order
+     */
+    private fun resetList() {
+        wordList = mutableListOf(
+            "queen",
+            "hospital",
+            "basketball",
+            "cat",
+            "change",
+            "snail",
+            "soup",
+            "calendar",
+            "sad",
+            "desk",
+            "guitar",
+            "home",
+            "railway",
+            "zebra",
+            "jelly",
+            "car",
+            "crow",
+            "trade",
+            "bag",
+            "roll",
+            "bubble"
+        )
+        wordList.shuffle()
+    }
+
+    /** Methods for buttons presses **/
+    fun onSkip() {
+        score--
+        nextWord()
+    }
+
+    fun onCorrect() {
+        score++
+        nextWord()
+    }
+
+    /**
+     * Moves to the next word in the list
+     */
+    private fun nextWord() {
+        if (!wordList.isEmpty()) {
+            //Select and remove a word from the list
+            word = wordList.removeAt(0)
+        }
+    }
+}
